Change int4 mapping to defatult Int, we still need to look for a better fit

+++ internal/provider/psql2ch_data_source.go
@@ -178,7 +178,7 @@ func postgreSqlToClickhouseType(psqlType string, numericPrecision int64, numeric
 	clickhouseType := ""
 	switch psqlType {
 	case "int4":
-		clickhouseType = "Int8"
+		clickhouseType = "Int"
 	case "numeric":
 		clickhouseType = fmt.Sprintf("Decimal(%d, %d)", numericPrecision, numericScale)
 	case "varchar":
